requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.0
pdf2docx==0.5.6
reportlab==4.0.7

# NLP and ML
spacy==3.7.2
nltk==3.8.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.4

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
structlog==23.2.0
rich==13.7.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional utilities
aiofiles==23.2.1
pillow==10.1.0
openpyxl==3.1.2
